pyproject.toml
[project]
name = "livrariaFabrica"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Luís Carlos Martins Junior", email = "luisc.martinsjunior5@gmail.com"},
]
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.1",
    "django-cors-headers>=4.3.1",
    "djangorestframework-simplejwt>=5.3.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
